Hide center clock option from the poor souls with notch devices
This is self-explainatory I think. Devices like the OnePlus 6/Pixel
3XL, oh and iPhone X have that big notch (for some reason) blocking
that nice big chunk of screen in the center of the status bar where
the good ol' center clock normally hung out. With such notch-cursed
devices in existence, we need to check for status_bar_height in order
to determine if the device has a notch or not.

It's hard enough I'm sure, they don't need another constant reminder
of what "could have been".

Signed-off-by: calebcabob <calphonic@gmail.com>